fixed prototype of GdiGetCharDimensions
svn path=/trunk/; revision=19841

+++ reactos/lib/gdi32/objects/font.c
@@ -818,7 +818,7 @@ RemoveFontResourceA(
  */
 DWORD
 STDCALL
-GdiGetCharDimensions(HDC hdc, LPTEXTMETRICW lptm, DWORD *height)
+GdiGetCharDimensions(HDC hdc, LPTEXTMETRICW lptm, LONG *height)
 {
     SIZE sz;
     static const WCHAR alphabet[] = {
